enfiproV2/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-dbbackup==4.2.1
et_xmlfile==2.0.0
idna==3.10
nodeenv==1.9.1
openpyxl==3.1.2
psycopg2-binary==2.9.10
pyserial==3.5
pytz==2025.2
pyusb==1.2.1
requests==2.31.0
sqlparse==0.5.3
urllib3==2.4.0
